so many fucking changes
-added more examples to the insults
-added a score tracker to insult fights
-added Macaque Island Treasure Coins

Coins are gained from insult fights and currently have no use, but will be used to buy drinks from the tavern... once I add the drinks

-added drink framework, and a few example drinks that currently do nothing

Next up, i'm going to be adding a drink that makes a crewmate lucky, a drink that cures a sick crewmate, and a drink that gives the player a new opening insult

-added a debug command to the port to give all insults
-added command to check coins
-fixed unintended crash when trying to show field description after fight
-implemented Sword Master boss fight

The fight works as intended, but it's a bit of a bitch to actually win if you get unlucky. Maybe having a Lucky crewmate should increase your odds, to give a use for the Lucky drink.