reorganized shopping
-added a Coinpurse to dialogue
-moved coins tracker into Coinpurse
-port debug now sets coins to 999
-added shop function to dialogue

Characters can now be used as shopkeepers right from the dialogue module. When you buy an item, the shop function takes your coins and returns the name of the item you bought. Does this mean I know how that item will actually work? No, of course not, but it does take your money