implemented drinks and boss luck
-changed literally ONE SINGULAR UNDERSCORE in dialogue
-added effects for all drinks

I had to completely redo how i did the Drinks class and make some of the sloppiest looking code i've ever had the misfortune of writing down to do it, but at least they all function as intended

-added luck mechanic to sword master fight

If you have a lucky crewmate, their luck with increase your chance of getting a lucky hit and standing your ground during the Sword Master boss fight. I may still mess around with the values a bit, but for now this is fine.