forgot some placeholders
so remember when i said i did ALL the dialogue and removed ALL the placeholders? yeah, i forgot the Cartographer and the Pirate names. technically there's one more Placeholder in the Sword Master's dialogue, but it's there on purpose and it's literally impossible for the player to ever see it.